refactor(Ladder,Point) : Point 도메인 ladder 패키지로 이동 , Direction 가져오는 Ladder 함수 매개변수 Point 로 변경
Point 가 Ladder 내에서 작동한다 고려하여 package 이동
getDirectionWithPoint 매개변수 row,column -> Point

- 이에 따른 전반적 import 문 수정

-[Command 관련 피드백 링크](https://github.com/woowacourse/java-ladder/pull/365#discussion_r1507542620)
